+++ src/ffvship_utility/CLI_Parser.hpp
@@ -21,7 +21,7 @@ struct ArgParser {
     std::vector<FlagGroup> flag_groups;
     std::map<std::string, int> alias_map; //returns index in flag_groups
     std::vector<int> positional_indexing;
-    int positional_counter = 0;
+    uint positional_counter = 0;
 
     ArgParser() {
         add_flag({"-h", "--help"}, &show_help_flag, "Display this help message");
@@ -157,7 +157,7 @@ private:
     void print_help() const {
         std::cout << "Usage : " << std::endl;
         std::cout << binary_name << " [";
-        for (int pos_id = 0; pos_id < positional_indexing.size(); pos_id++){
+        for (unsigned long pos_id = 0; pos_id < positional_indexing.size(); pos_id++){
             std::cout << "Positional " << pos_id;
             auto& group = flag_groups[positional_indexing[pos_id]];
             if (group.aliases.size() > 0){
